Create card and section files
